// packages/functions/src/types.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type TheTypeSchema = string;
export type TheTypeSchema1 = string;
export type TheTypeSchema2 = string;
export type ASchema3 = number;
export type ASchema2 = ASchema3[];
export type ASchema1 = ASchema2[];
export type TheCoordinatesSchema = ASchema1[];
export type TheFeaturesSchema = ASchema[];
export type TheTypeSchema3 = string;
export type TheTypeSchema4 = string;
export type TheTypeSchema5 = string;
export type ASchema7 = number;
export type ASchema6 = ASchema7[];
export type ASchema5 = ASchema6[];
export type TheCoordinatesSchema1 = ASchema5[];
export type TheElevationSchema = number;
export type TheFeaturesSchema1 = ASchema4[];

export interface RootSchema {
  building_limits: TheBuildingLimitsSchema;
  height_plateaus: TheHeightPlateausSchema;
  [k: string]: unknown;
}
export interface TheBuildingLimitsSchema {
  type: TheTypeSchema;
  features: TheFeaturesSchema;
  [k: string]: unknown;
}
export interface ASchema {
  type: TheTypeSchema1;
  properties: ThePropertiesSchema;
  geometry: TheGeometrySchema;
  [k: string]: unknown;
}
export interface ThePropertiesSchema {
  [k: string]: unknown;
}
export interface TheGeometrySchema {
  type: TheTypeSchema2;
  coordinates: TheCoordinatesSchema;
  [k: string]: unknown;
}
export interface TheHeightPlateausSchema {
  type: TheTypeSchema3;
  features: TheFeaturesSchema1;
  [k: string]: unknown;
}
export interface ASchema4 {
  type: TheTypeSchema4;
  geometry: TheGeometrySchema1;
  properties: ThePropertiesSchema1;
  [k: string]: unknown;
}
export interface TheGeometrySchema1 {
  type: TheTypeSchema5;
  coordinates: TheCoordinatesSchema1;
  [k: string]: unknown;
}
export interface ThePropertiesSchema1 {
  elevation: TheElevationSchema;
  [k: string]: unknown;
}